requirements.txt
boto==2.49.0
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dialogflow==0.5.2
Flask==1.0.2
future==0.17.1
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.19.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
line-bot-sdk==1.8.0
MarkupSafe==1.1.1
protobuf==3.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
